add three main methods to manage ticket and token
